requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jax==0.4.28
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.12.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
ml-dtypes==0.4.0
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
